+++ utils/constants.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-APP_VERSION = "0.2.123"
+APP_VERSION = "0.2.124"
 
 
 # Player tags